# backend/openapi.json
{
  "components": {
    "schemas": {
      "BotanicalIdentityCreate": {
        "additionalProperties": false,
        "properties": {
          "common_name": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Common Name"
          },
          "cultivar_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cultivar Name"
          },
          "scientific_name": {
            "maxLength": 255,
            "title": "Scientific Name",
            "type": "string"
          }
        },
        "required": [
          "scientific_name"
        ],
        "title": "BotanicalIdentityCreate",
        "type": "object"
      },
      "BotanicalIdentityResponse": {
        "properties": {
          "common_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Common Name"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cultivar_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cultivar Name"
          },
          "display_label": {
            "title": "Display Label",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "scientific_name": {
            "title": "Scientific Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scientific_name",
          "cultivar_name",
          "common_name",
          "created_at",
          "updated_at",
          "display_label"
        ],
        "title": "BotanicalIdentityResponse",
        "type": "object"
      },
...
      "BotanicalProfilePut": {
        "additionalProperties": false,
        "properties": {
          "cultivation": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cultivation"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "origin_distribution": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Distribution"
          },
          "uses": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uses"
          },
          "warnings": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warnings"
          }
        },
        "title": "BotanicalProfilePut",
        "type": "object"
      },
      "BotanicalProfileResponse": {
        "properties": {
          "botanical_identity_id": {
            "format": "uuid",
            "title": "Botanical Identity Id",
            "type": "string"
          },
          "cultivation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cultivation"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "origin_distribution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Distribution"
          },
          "uses": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uses"
          },
          "warnings": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warnings"
          }
        },
        "required": [
          "botanical_identity_id",
          "description",
          "origin_distribution",
          "cultivation",
          "uses",
          "warnings"
        ],
        "title": "BotanicalProfileResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Florabase API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/botanical-identities": {
      "get": {
        "operationId": "listBotanicalIdentities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BotanicalIdentityResponse"
                  },
                  "title": "Response Listbotanicalidentities",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "summary": "List All",
        "tags": [
          "botanical identities"
        ]
      },
      "post": {
        "operationId": "createBotanicalIdentity",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BotanicalIdentityCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BotanicalIdentityResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Location": {
                "description": "Canonical URL of the created BotanicalIdentity",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Request forbidden"
          },
          "409": {
            "description": "Botanical identity already exists"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "botanical identities"
        ]
      }
    },
    "/api/v1/botanical-identities/{botanical_identity_id}": {
      "get": {
        "operationId": "getBotanicalIdentity",
        "parameters": [
          {
            "in": "path",
            "name": "botanical_identity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Botanical Identity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BotanicalIdentityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Botanical identity not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read",
        "tags": [
          "botanical identities"
        ]
      }
    },
    "/api/v1/botanical-identities/{botanical_identity_id}/profile": {
      "get": {
        "operationId": "getBotanicalProfile",
        "parameters": [
          {
            "in": "path",
            "name": "botanical_identity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Botanical Identity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BotanicalProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Identity or profile not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read",
        "tags": [
          "botanical profiles"
        ]
      },
      "put": {
        "operationId": "putBotanicalProfile",
        "parameters": [
          {
            "in": "path",
            "name": "botanical_identity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Botanical Identity Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BotanicalProfilePut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BotanicalProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "201": {
            "description": "Botanical profile created"
          },
          "204": {
            "description": "Final section cleared; profile removed"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Request forbidden"
          },
          "404": {
            "description": "Botanical identity not found"
          },
          "422": {
            "description": "Profile content invalid"
          }
        },
        "summary": "Put",
        "tags": [
          "botanical profiles"
        ]
      }
    },
...
  }
}